+++ lib/screens/Lobby/created_game_tile.dart
@@ -1,83 +1,6 @@
-import 'dart:html';
-
 import 'package:flutter/material.dart';
 import 'package:niira/models/game.dart';
 import 'package:niira/screens/input_password.dart';
-import 'package:niira/services/auth/auth_service.dart';
-import 'package:niira/services/database/database_service.dart';
-import 'package:niira/services/location_service.dart';
-import 'package:niira/navigation/navigation.dart';
-import 'package:provider/provider.dart';
-
-class LobbyScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Lobby'),
-        actions: [
-          FlatButton(
-            key: Key('signOutBtn'),
-            onPressed: () {
-              // create a function to call on confirmation
-              final signOut = () async {
-                Navigator.of(context).pop();
-                await context.read<AuthService>().signOut();
-              };
-
-              context.read<Navigation>().showConfirmationDialog(
-                    onConfirmed: signOut,
-                    confirmText: 'Sign Out',
-                    cancelText: 'Return',
-                  );
-            },
-            child: Text('log out'),
-          )
-        ],
-      ),
-      body: Container(
-        child: ListOfCreatedGames(),
-      ),
-    );
-  }
-}
-
-class ListOfCreatedGames extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    final locationService = context.read<LocationService>();
-    final userLocation = await locationService
-        .getUsersCurrentLocation; // is async, convert to stateful widget?
-    final createdGamesWithDistance = context
-        .watch<DatabaseService>()
-        .streamOfCreatedGames
-        .map<List<Game>>((games) => locationService
-            .setDistanceBetweenUserAndGames(games, userLocation));
-
-    final createdGamesInOrderOfDistance = createdGamesWithDistance.map(
-        (_games) => _games
-            .sort((a, b) => a.distanceFromUser.compareTo(b.distanceFromUser)));
-    return StreamBuilder<List<Game>>(
-        stream: createdGamesWithDistance,
-        builder: (context, snapshot) {
-          if (snapshot.data == null) {
-            return Container();
-          } else {
-            return Padding(
-              padding: EdgeInsets.fromLTRB(0, 10, 0, 0),
-              child: ListView.builder(
-                itemCount: snapshot.data.length,
-                itemBuilder: (context, index) {
-                  return GameTile(
-                    snapshot.data[index],
-                  );
-                },
-              ),
-            );
-          }
-        });
-  }
-}
 
 class GameTile extends StatelessWidget {
   final Game _game;
@@ -103,7 +26,7 @@ class GameTile extends StatelessWidget {
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: <Widget>[
                       Text(
-                        '500m away', // TODO: will replace with real data in ticket #60
+                        '${_game.distanceFromUser}m away',
                         textAlign: TextAlign.left,
                         style: TextStyle(
                           fontSize: 14,
