+++ test/create_game_section/Lobby/lobby_widget_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:niira/main.dart';
 import 'package:niira/models/game.dart';
 import 'package:niira/models/user_data.dart';
-import 'package:niira/screens/lobby.dart';
+import 'package:niira/screens/Lobby/lobby.dart';
 import 'package:niira/services/database/database_service.dart';
 import 'package:niira/navigation/navigation.dart';
 import 'package:provider/provider.dart';
